Added Component Lifecycle Methods
In order to demonstrate the component lifecycle methods available in
React, we added the two most popular, componentDidMount() and
componentDidUpdate() with console.logs to show that these functions are
called whenever the app loads up, renders, and rerenders data to the DOM

- componentDidMount() is called whenever the app initially loads
- componentDidUpdate() is called whenever manipulation to the state
occurs and the render method is called again. In the case of this
trivial app, componentDidMount() is called when the page is loaded for
the first time (or whenever we reset the location request back to Ask)
and componentDidUpdate is called whenever the location request is
answered

These functions are called by default when the app loads, but we can
make them explicit and also utilize them to handle data loading
and manipulation.